feat: Parse.MessageIterator.unpack
unpack an OSC message into a tuple type determined by the type tags
expected. returns error.TypeMismatch or error.InvalidOSC on failure,
and a tuple of the specified argument types on success.

basic type polymorphism is supported, for example "!?i" will match an
argument of type I, N or i and return a value of type `error{Bang}!?i32`.